v2/requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
requests>=2.28.0
rich>=13.0.0
